+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [3.0.0] - 2019-09-17
+### Changed
+- Plugin classes moved to wp-builder
+### Added
+- Factory can create checker from requirement array
+- Support for update suggestion when required plugin not found
+
 ## [2.4.0] - 2019-06-04
 ### Added
 - Plugin name in plugin info